crypto/enable_fips: remove invalid fips-mode-setup package from dependencies
Utility fips-mode-setup is provided by package crypto-policies-scripts.
The fact that fips-mode-setup is declared as a dependency is generating
trouble on RHEL 8 jobs, as restraint understands it as a package and will
generate an error when it is not able to find it. This commit simply
removes the tool from both dependency definitions, so that the issue
no longer happens.