perf tests: parse results with test ID after 9.4
Test results in 9.0 or 9.4 do not exhibit test id prefix in the result
line. Such as [1]:

	---- end ----
	Test data source output: Ok

	---- end ----
	Zstd perf.data compression/decompression: Ok

Adjust the rhel version check to account for this, since not every RHEL9
stream is showing the test ID prefix in the result line.

Fall back to the previous parsing method when the detected RHEL version
is 9.4 or before.

Fixes: 8d640c46fe10cb75 ("perf tests: Fix parsing of perf test result
without test ID prefix")

[1]: https://beaker.engineering.redhat.com/recipes/19355151#task199751373

Signed-off-by: Jose Guisado <jguisado@redhat.com>