crypto/enable_fips: Move crypto-policies-scripts as soft dependency
There is no crypto-policies-scripts package on 8.2z. fips-mode-setup
is provided instead by the crypto-policies package. This commit
moves crypto-policies-scripts then to a soft dependency, one that will
get installed if available and skipped if not. This will fix restraint
aborting the test on 8.2z, while keeping the other streams same as
always.