kselftests/net: fix timeout in fib_test
The net selftest always times out on c9s. This happens because fib_test
forgets to source lib.sh, which causes kill_process to fail. As a result,
some background processes remain running and block the test from finishing.

Fix this temporarily by sourcing lib.sh in the script, until a proper
solution is implemented in the kernel.

Signed-off-by: Hangbin Liu <haliu@redhat.com>