[rt] remove the 100 forks testing from rt_ssdd
ssdd is a regression test to validate the reliability of PTRACE_SINGLESTEP, the
defaults 10 forks is enough to trigger the original issue, increasing the forks
will add overheads to the system which in some cases causes the program to fail.
This commit removes the 100 forks scenario as it adds little value regarding
triggering the initial problem but instead adds more uncertainty.

For more discussions on the issue caused by the extra forks, see RHEL-13213.

Signed-off-by: Shizhao Chen <shichen@redhat.com>