memory/function/memeq: print MEM (size list)
A issue hit that in 6T mmemory box, the mem=4G is tested, and what's
worse is, it's left in the command line in the whole beaker job, without
cleaning up. So add this to see what happens, 4G shouldn't in the test
list, and the parameter should be cleanuped after the test finish.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>