memory/function/memeq: fix mem size list setup
set_mem() set the MEM variable which is a list of all memory sizes that
need to be tested. It's flagged by the tmp dir so that it only runs once,
when test begins and no any mem= parameter is setup.  if this flag is
unreliable or disappered, the test will run set_mem() in every reboot,
and the total memory and used memory info changes every reboot, so
the size list (MEM) parameter changes every reboot, as a result, every
memory size is tested, including the trouble (mem=4G), which should be
run only for boxes whose physical memory less than 16G host.

The reason why the flag dir disappered after reboot is the OUTPUTFILE
variable is a tmp dir with mktemp:
```
Setting OUTPUTFILE as /tmp/tmp.NmJ62TvG1o/resultoutputfile.log
```

This issue causes several issues:
1) kernel panic on memory deadlock on kernel-debug x86_64, where only 1.4G
   total memory left for testing and can't boot.

2) causes 6T machine kernel panic in VMM-FUNCTION test, as there's the
   no flag dir exist, so the test finally existed with mem=4G in command
   line without cleanup:
   """
   Sorry, the system RAM is too low to test.
   """

So we choose to use a fixed dirname as the flag dir name.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>