security/crypto: fix LTP tag for different RHEL versions
To prevent issues with mismatch between a RHEL version and what is being
tested/compiled during LTP test we need to fix the version that is most
compatible and that was already tested as supported in the past per-release.
Using the distribution/ltp test component as a reference, this commit set
the LTP_TAG variable to a fixed number comparing the RHEL release the test
is running on.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>