rt-tests/printk: ignore the status of the kill/ps remaining pids
It's possible that the remaining processes were killed and cleaned up right
after they were acquired from cgroup.procs, in that case the ps and kill
commands will fail for no such process. This is a trivial problem as we only
care about whether can remove the root cgroup successfully, as to how and when
we do not care. So let's run the ps & kill in logonly mode.

Also extends the sleep time in cleanup, it'll take more than 10s for all
processes to quit.

Resovles: RHELTEST-1817

Signed-off-by: Shizhao Chen <shichen@redhat.com>